avatar/requirements.txt
# Avatar Generation Dependencies
# ================================

# Core
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Face Analysis & Swap
insightface>=0.7.0
onnxruntime-gpu>=1.12.0  # Use onnxruntime for CPU-only

# Face Enhancement (optional but recommended)
gfpgan>=1.3.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Video Processing
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0
